assumption before error
